ReadLoadingConfig, BamReaderAPI to own files
