+++ src/main/scala/org/hammerlab/guacamole/Common.scala
@@ -38,7 +38,8 @@ import org.hammerlab.guacamole.Concordance.ConcordanceArgs
 import org.hammerlab.guacamole.distributed.LociPartitionUtils
 import org.hammerlab.guacamole.loci.LociSet
 import org.hammerlab.guacamole.logging.LoggingUtils.progress
-import org.hammerlab.guacamole.reads.Read
+import org.hammerlab.guacamole.reads.BamReaderAPI.BamReaderAPI
+import org.hammerlab.guacamole.reads.{BamReaderAPI, Read, ReadLoadingConfig}
 import org.kohsuke.args4j.{Option => Args4jOption}
 
 /**
@@ -71,7 +72,7 @@ object Common extends Logging {
       var noSequenceDictionary: Boolean = false
     }
 
-    /** Argument for configuring read loading with a Read.ReadLoadingConfig object. */
+    /** Argument for configuring read loading with a ReadLoadingConfig object. */
     trait ReadLoadingConfigArgs extends Base {
       @Args4jOption(name = "--bam-reader-api",
         usage = "API to use for reading BAMs, one of: best (use samtools if file local), samtools, hadoopbam")
@@ -79,9 +80,8 @@ object Common extends Logging {
     }
     object ReadLoadingConfigArgs {
       /** Given commandline arguments, return a ReadLoadingConfig. */
-      def fromArguments(args: ReadLoadingConfigArgs): Read.ReadLoadingConfig = {
-        Read.ReadLoadingConfig(
-          bamReaderAPI = Read.ReadLoadingConfig.BamReaderAPI.withNameCaseInsensitive(args.bamReaderAPI))
+      def fromArguments(args: ReadLoadingConfigArgs): ReadLoadingConfig = {
+        ReadLoadingConfig(BamReaderAPI.withNameCaseInsensitive(args.bamReaderAPI))
       }
     }
