+++ src/main/scala/org/hammerlab/guacamole/ReadSet.scala
@@ -21,7 +21,7 @@ package org.hammerlab.guacamole
 import org.apache.spark.SparkContext
 import org.apache.spark.rdd.RDD
 import org.bdgenomics.adam.models.SequenceDictionary
-import org.hammerlab.guacamole.reads.{ MappedRead, PairedRead, Read }
+import org.hammerlab.guacamole.reads.{MappedRead, PairedRead, Read, ReadLoadingConfig}
 
 /**
  * A ReadSet contains an RDD of reads along with some metadata about them.
@@ -93,7 +93,7 @@ object ReadSet {
     filename: String,
     filters: Read.InputFilters = Read.InputFilters.empty,
     contigLengthsFromDictionary: Boolean = true,
-    config: Read.ReadLoadingConfig = Read.ReadLoadingConfig.default): ReadSet = {
+    config: ReadLoadingConfig = ReadLoadingConfig.default): ReadSet = {
 
     val (reads, sequenceDictionary) =
       Read.loadReadRDDAndSequenceDictionary(
